+++ refinedstorage2-network-api/src/main/java/com/refinedmods/refinedstorage2/api/network/component/StorageNetworkComponent.java
@@ -18,7 +18,9 @@ public class StorageNetworkComponent implements NetworkComponent {
 
     private final Map<StorageChannelType<?>, StorageChannel<?>> channels = new HashMap<>();
 
-    public StorageNetworkComponent(final OrderedRegistry<?, StorageChannelType<?>> storageChannelTypeRegistry) {
+    public StorageNetworkComponent(
+        final OrderedRegistry<?, ? extends StorageChannelType<?>> storageChannelTypeRegistry
+    ) {
         for (final StorageChannelType<?> type : storageChannelTypeRegistry.getAll()) {
             channels.put(type, type.create());
         }